requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
xgboost>=1.5.0
plotly>=5.0.0
jupyter>=1.0.0
pytest>=6.0.0
pyyaml>=5.4.0